+++ src/backend/chat/tests/views/chat/conversations/test_conversation_with_image_url.py
@@ -211,6 +211,7 @@ def test_post_conversation_with_local_image_url(
                 },
             ],
             "run_id": _run_id,
+            "state": "complete",
             "timestamp": "2025-10-18T20:48:20.286204Z",
         },
         {
@@ -794,6 +795,7 @@ def test_post_conversation_with_local_image_url_in_history(
                 }
             ],
             "run_id": _run_id,
+            "state": "complete",
             "timestamp": "2025-10-18T20:48:20.286204Z",
         },
         {